requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
aom=3.6.1=h59595ed_0
asttokens=3.0.0=pyhd8ed1ab_1
blas=1.0=mkl
blosc=1.21.6=hef167b5_0
brotli=1.1.0=hb9d3cd8_2
brotli-bin=1.1.0=hb9d3cd8_2
brotli-python=1.1.0=py39hf88036b_2
brunsli=0.1=h9c3ff4c_0
bzip2=1.0.8=h4bc722e_7
c-blosc2=2.14.4=hb4ffafa_1
ca-certificates=2025.4.26=hbd8a1cb_0
certifi=2025.1.31=pyhd8ed1ab_0
cffi=1.17.1=py39h15c3d72_0
charls=2.4.2=h59595ed_0
charset-normalizer=3.4.1=pyhd8ed1ab_0
comm=0.2.2=pyhd8ed1ab_1
contourpy=1.3.0=pypi_0
cpython=3.9.21=py39hd8ed1ab_1
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cycler=0.12.1=pypi_0
dav1d=1.2.1=hd590300_0
debugpy=1.8.14=py39hf88036b_0
decorator=5.2.1=pyhd8ed1ab_0
einops=0.8.1=pypi_0
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.2.0=pyhd8ed1ab_0
ffmpeg=4.4.0=hca11adc_0
filelock=3.16.1=pyhd8ed1ab_1
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_3
fontconfig=2.15.0=h7e30c49_1
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.55.6=pypi_0
freetype=2.12.1=h267a509_2
fsspec=2024.12.0=pypi_0
future=1.0.0=pypi_0
gettext=0.22.5=he02047a_3
gettext-tools=0.22.5=he02047a_3
giflib=5.2.2=hd590300_0
gmp=6.3.0=hac33072_2
gmpy2=2.1.5=py39h7196dd7_3
gnutls=3.6.13=h85f3911_1
h2=4.1.0=pyhd8ed1ab_1
hpack=4.0.0=pyhd8ed1ab_1
hyperframe=6.0.1=pyhd8ed1ab_1
idna=3.10=pyhd8ed1ab_1
imagecodecs=2024.1.1=py39ha98d97a_6
imageio=2.37.0=pyhfb79c49_0
importlib-metadata=8.6.1=pyha770c72_0
importlib-resources=6.5.2=pypi_0
intel-openmp=2022.0.1=h06a4308_3633
ipykernel=6.29.5=pyh3099207_0
ipython=8.18.1=pyh707e725_3
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.4=pypi_0
joblib=1.4.2=pyhd8ed1ab_1
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.7.2=pyh31011fe_1
jxrlib=1.1=hd590300_3
kaleido=0.2.1=pypi_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.7=pypi_0
krb5=1.21.3=h659f571_0
lame=3.100=h166bdaf_1003
lazy-loader=0.4=pyhd8ed1ab_2
lazy_loader=0.4=pyhd8ed1ab_2
lcms2=2.16=hb7c19ff_0
ld_impl_linux-64=2.43=h712a8e2_2
lerc=4.0.0=h27087fc_0
libaec=1.1.3=h59595ed_0
libasprintf=0.22.5=he8f35ee_3
libasprintf-devel=0.22.5=he8f35ee_3
libavif16=1.0.1=h87da1f6_2
libblas=3.9.0=16_linux64_mkl
libbrotlicommon=1.1.0=hb9d3cd8_2
libbrotlidec=1.1.0=hb9d3cd8_2
libbrotlienc=1.1.0=hb9d3cd8_2
libcblas=3.9.0=16_linux64_mkl
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.9.1.3=0
libcurand=10.3.5.147=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.20=hd590300_0
libdrm=2.4.124=hb9d3cd8_0
libedit=3.1.20250104=pl5321h7949ede_0
libegl=1.7.0=ha4b6fd6_2
libexpat=2.6.4=h5888daf_0
libffi=3.4.2=h7f98852_5
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgettextpo=0.22.5=he02047a_3
libgettextpo-devel=0.22.5=he02047a_3
libgfortran=14.2.0=h69a702a_1
libgfortran-ng=14.2.0=h69a702a_1
libgfortran5=14.2.0=hd5240d6_1
libgl=1.7.0=ha4b6fd6_2
libglvnd=1.7.0=ha4b6fd6_2
libglx=1.7.0=ha4b6fd6_2
libgomp=14.2.0=h77fa898_1
libhwy=1.1.0=h00ab1b0_0
libiconv=1.17=hd590300_2
libidn2=2.3.7=hd590300_0
libjpeg-turbo=3.0.0=hd590300_1
libjxl=0.10.3=h66b40c8_0
liblapack=3.9.0=16_linux64_mkl
liblzma=5.6.3=hb9d3cd8_1
liblzma-devel=5.6.3=hb9d3cd8_1
libnpp=11.8.0.86=0
libnsl=2.0.1=hd590300_0
libnvjpeg=11.9.0.86=0
libopenblas=0.3.28=pthreads_h94d23a6_1
libpciaccess=0.18=hd590300_0
libpng=1.6.44=hadc24fc_0
libsodium=1.0.20=h4ab18f5_0
libsqlite=3.47.2=hee588c1_0
libstdcxx=14.2.0=hc0a3c3a_1
libstdcxx-ng=14.2.0=h4852527_1
libtasn1=4.19.0=h166bdaf_0
libtiff=4.6.0=h1dd3fc0_3
libunistring=0.9.10=h7f98852_0
libuuid=2.38.1=h0b41bf4_0
libva=2.22.0=h8a09558_1
libvpx=1.13.1=h59595ed_0
libwebp=1.4.0=h2c329e2_0
libwebp-base=1.4.0=hd590300_0
libxcb=1.17.0=h8a09558_0
libxcrypt=4.4.36=hd590300_1
libxml2=2.13.5=h0d44e9d_1
libzlib=1.3.1=hb9d3cd8_2
libzopfli=1.0.3=h9c3ff4c_0
lightning-utilities=0.12.0=pypi_0
llvm-openmp=15.0.7=h0cdce71_0
lpips=0.1.4=pypi_0
lz4-c=1.9.4=hcb278e6_0
markupsafe=3.0.2=py39h9399b63_1
matplotlib=3.9.4=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mkl=2022.1.0=hc2b9512_224
mpc=1.3.1=h24ddda3_1
mpfr=4.2.1=h90cbb55_3
mpmath=1.3.0=pyhd8ed1ab_1
ncurses=6.5=he02047a_1
nest-asyncio=1.6.0=pyhd8ed1ab_1
nettle=3.6=he412f7d_0
networkx=3.2.1=pyhd8ed1ab_0
numpy=1.26.4=py39h474f0d3_0
nvidia-cublas-cu12=12.4.5.8=pypi_0
nvidia-cuda-cupti-cu12=12.4.127=pypi_0
nvidia-cuda-nvrtc-cu12=12.4.127=pypi_0
nvidia-cuda-runtime-cu12=12.4.127=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.2.1.3=pypi_0
nvidia-curand-cu12=10.3.5.147=pypi_0
nvidia-cusolver-cu12=11.6.1.9=pypi_0
nvidia-cusparse-cu12=12.3.1.170=pypi_0
nvidia-nccl-cu12=2.21.5=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.4.127=pypi_0
odak=0.2.7=dev_0
opencv-python=4.10.0.84=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=h488ebb8_0
openssl=3.5.0=h7b32b05_0
p11-kit=0.24.1=hc5aa10d_0
packaging=24.2=pyhd8ed1ab_2
parso=0.8.4=pyhd8ed1ab_1
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.0.0=pypi_0
pip=24.3.1=pyh8b19718_2
platformdirs=4.3.7=pyh29332c3_0
plotly=5.24.1=pypi_0
plyfile=1.1=pypi_0
prompt-toolkit=3.0.51=pyha770c72_0
psutil=7.0.0=py39h8cd3c5a_0
pthread-stubs=0.4=hb9d3cd8_1002
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pycparser=2.22=pyh29332c3_1
pygments=2.19.1=pyhd8ed1ab_0
pyparsing=3.2.1=pypi_0
pysocks=1.7.1=pyha55dd90_7
python=3.9.21=h9c0c6dc_1_cpython
python-dateutil=2.9.0.post0=pyhff2d567_1
python_abi=3.9=5_cp39
pytorch=2.5.1=py3.9_cuda11.8_cudnn9.1.0_0
pytorch-cuda=11.8=h7e8668a_6
pytorch-mutex=1.0=cuda
pywavelets=1.6.0=py39hd92a3bb_0
pyyaml=6.0.2=py39h8cd3c5a_1
pyzmq=26.4.0=py39h4e4fb57_0
rav1e=0.6.6=he8a937b_2
readline=8.2=h8228510_1
regex=2024.11.6=pypi_0
requests=2.32.3=pyhd8ed1ab_1
safetensors=0.5.2=pypi_0
scikit-image=0.24.0=py39h3b40f6f_3
scikit-learn=1.6.1=py39h4b7350c_0
scipy=1.13.1=py39haf93ffa_0
setuptools=75.6.0=pyhff2d567_1
six=1.17.0=pyhd8ed1ab_0
snappy=1.2.1=h8bd8927_1
stack_data=0.6.3=pyhd8ed1ab_1
svt-av1=1.7.0=h59595ed_0
sympy=1.13.1=pypi_0
tenacity=9.0.0=pypi_0
threadpoolctl=3.5.0=pyhc1e730c_0
tifffile=2024.6.18=pyhd8ed1ab_0
tk=8.6.13=noxft_h4845f30_101
tokenizers=0.21.0=pypi_0
torch=2.5.1=pypi_0
torchaudio=2.5.1=py39_cu118
torchmetrics=1.6.1=pypi_0
torchtriton=3.1.0=py39
torchvision=0.20.1=py39_cu118
tornado=6.4.2=py39h8cd3c5a_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pyhd8ed1ab_1
transformers=4.48.3=pypi_0
triton=3.1.0=pypi_0
typing_extensions=4.12.2=pyha770c72_1
tzdata=2024b=hc8b5060_0
urllib3=2.3.0=pyhd8ed1ab_0
wayland=1.23.1=h3e06ad9_0
wayland-protocols=1.37=hd8ed1ab_0
wcwidth=0.2.13=pyhd8ed1ab_1
wheel=0.45.1=pyhd8ed1ab_1
x264=1!161.3030=h7f98852_1
x265=3.5=h924138e_3
xorg-libx11=1.8.10=h4f16b4b_1
xorg-libxau=1.0.12=hb9d3cd8_0
xorg-libxdmcp=1.1.5=hb9d3cd8_0
xorg-libxext=1.3.6=hb9d3cd8_0
xorg-libxfixes=6.0.1=hb9d3cd8_0
xz=5.6.3=hbcc6ac9_1
xz-gpl-tools=5.6.3=hbcc6ac9_1
xz-tools=5.6.3=hb9d3cd8_1
yaml=0.2.5=h7f98852_2
zeromq=4.3.5=h3b0a872_7
zfp=1.0.1=h5888daf_2
zipp=3.21.0=pyhd8ed1ab_1
zlib-ng=2.0.7=h0b41bf4_0
zstandard=0.23.0=py39h08a7858_1
zstd=1.5.6=ha6fb4c9_0
